requirements.txt
Flask
python-dotenv
langgraph
langchain-core
langchain-google-genai
langchain-text-splitters
langchain-chroma
psycopg2-binary
chromadb